scrum-22: rename project folder
