--- NSU_Sports_Complex_System/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'NSU_Sports_Complex_System'
