--- NSU_Sports_Complex_System/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=NSU_Sports_Complex_System
