+++ backend/build.gradle
@@ -4,7 +4,6 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
-group = 'ru.nsu.fit'
 version = '0.0.1-SNAPSHOT'
 
 java {
