+++ backend/src/test/java/ru/nsu/sports/complex/backend/NsuSportsComplexSystemApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.nsu.fit.nsu_sports_complex_system;
+package ru.nsu.sports.complex.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
